Updated Amz Link, Template From RGBW -> RGBCCT
The old Amazon link went to a defunct listing.  Moved link2 -> link; removed link2.

The bulb type is listed here as RGBCCT but the template was for an RGBW.  It was very difficult to make HA switch from color_mode: "color_temp" -> color_mode: "hs".  The updated template makes it function like this template:

https://templates.blakadder.com/aoycocr_Q9WM.html

Bought a 4 pack on 8/17/21.  It would not flash OTA, but soldering connections to the updated board was quite easy.  I flashed them to Tasmota 9.5.0 with no issues.